requirements.txt
# Core ML/NLP Libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
tokenizers>=0.13.0
accelerate>=0.20.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Evaluation and Metrics
evaluate>=0.4.0
datasets[audio]>=2.12.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Web Interface
streamlit>=1.25.0
streamlit-option-menu>=0.3.6

# Configuration and Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
click>=8.1.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Jupyter and Documentation
jupyter>=1.0.0
notebook>=6.5.0
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Optional: GPU support (uncomment if needed)
# torch-audio>=2.0.0
# torch-vision>=0.15.0

# Optional: Advanced NLP features
# spacy>=3.6.0
# nltk>=3.8.0
# textstat>=0.7.0
